2nd Earth map, prevent player from leaving screen, restartgame adjusted, new Venus bg

+++ desktop/src/sceneManager/EarthScene.java
@@ -31,6 +31,7 @@ public class EarthScene extends BasePlanetScene{
 	private Entity flag;
 	private boolean showDialogFlag = true;
 	private boolean dialogOpen = false;
+	private boolean isSecondMapLoaded = false;
 
 	public EarthScene(SceneManager sceneManager, EntityManager entityManager, EntityFactory entityFactory, PlayerControllerManager pcManager,
             CollisionManager cManager, AIControlManager aiManager) {
@@ -49,12 +50,12 @@ public class EarthScene extends BasePlanetScene{
 		 /* set player in SceneManager to this created player entity. This allows other
 		 scenes to retrieve this particular player entity using getPlayer */
 		sceneManager.setPlayer(player);
-		
+
 		// flag entity
-		flag = entityFactory.createEntity("flag.png", 600, 350, false, "flag");
+		//flag = entityFactory.createEntity("flag.png", 600, 350, false, "flag");
 
         String buttonText = "End";
-        addButton(buttonText, Gdx.graphics.getWidth() - 100, Gdx.graphics.getHeight() - 50,
+        addButton(buttonText, Gdx.graphics.getWidth() - 125, Gdx.graphics.getHeight() - 50,
                 () -> sceneManager.showEndScene());
     }
 
@@ -73,6 +74,34 @@ public class EarthScene extends BasePlanetScene{
         showCustomDialog("", "Earth is our home planet and uniquely"
         		+ "\nsupports life due to water and protective atmosphere.", windowStyle);
     }
+
+	// check which Earth map player is on
+	public boolean isSecondMapLoaded() {
+		return isSecondMapLoaded;
+	}
+
+	// for player to load back into first map when colliding with lava in second map
+	public void loadFirstMap() {
+		mapManager.loadMap("Earth.tmx");
+		isSecondMapLoaded = false;
+	}
+
+	// switch to second Earth map when player exits first Earth map
+	public void switchMap() {
+		if (player.getPosX() > 620) {
+			// Load the second map if it's not already loaded
+			if (!isSecondMapLoaded) {
+				mapManager.loadMap("Earth2.tmx"); // Load the second map
+				isSecondMapLoaded = true; // Set the flag to true
+				// Create the flag entity only when Earth2.tmx is loaded and the flag entity does not exist yet
+				if (flag == null) {
+					flag = entityFactory.createEntity("flag.png", 570, 400, false, "flag");
+					entityManager.add(flag);
+					entityManager.addCollidableEntity(flag);
+				}
+			}
+		}
+	}
 	
 	@Override
 	public void render(float delta) {
@@ -85,6 +114,7 @@ public class EarthScene extends BasePlanetScene{
 		mapManager.getRenderer().render();
 		// Check collision with building
 		cManager.checkCollisionWithObject((Player) player, mapManager);
+		switchMap(); // call map switching function when player exits bounds of first Earth map
 
 	    SpriteBatch batch = new SpriteBatch();
 	    
@@ -105,19 +135,11 @@ public class EarthScene extends BasePlanetScene{
 	}
 
 	@Override
-	public void show() {
-		super.show();
-		if (flag != null) {
-			entityManager.add(flag);
-			entityManager.addCollidableEntity(flag);
-		}
-	}
-
-	@Override
-	public void hide() {
+	public void hide() { // to remove flag from first Earth map
 		super.hide();
-		if (flag != null) {
+		if (flag != null && !isSecondMapLoaded) {
 			entityManager.removeFlagEntity(flag);
+			flag = null; // Set flag to null after removing it
 		}
 	}
 }
